pyproject.toml
[project]
name = "cloudappsproject1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
    "PyJWT>=2.8.0",
    "bcrypt>=4.1.0",
    "flask-mail>=0.10.0",
    "gunicorn>=25.0.1",
    "dnspython>=2.8.0",
    "grpcio-tools>=1.78.0",
    "grpcio>=1.78.0",
    "opentelemetry-distro>=0.51b0",
    "opentelemetry-exporter-otlp>=1.30.0",
    "opentelemetry-instrumentation-flask>=0.51b0",
    "opentelemetry-instrumentation-requests>=0.51b0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
